pyproject.toml
[project]
name = "EvoGames"
version = "0.2.0"
description = "Multi-agents simulation environment for evoluationary games."
requires-python = ">=3.12"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.5",
    "openai>=1.77.0",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
]
